Draft/Mockup of Project
I uploaded a link to a draft of what I want the final project to look like. I have not completed it, I just laid out on what I want the final to look like with some notes as to what I will be adding. Please let me now if I missed something.